+++ CHANGELOG.md
@@ -12,6 +12,9 @@
   only joining tables that are actually needed.
 - Added a `count_query` option to override the count query used by
   `Flop.run/3`, `Flop.validate_and_run/3` and `Flop.validate_and_run!/3`.
+- You can get a list of allowed operators for a given Ecto type or a given
+  schema field with `Flop.Filter.allowed_operators/1` and
+  `Flop.Filter.allowed_operators/2` now.
 
 ### Changed
 
@@ -23,6 +26,7 @@
   With this change, the configuration format is the same everywhere. A compile
   time exception is raised if you are still using the old format, guiding you in
   the update.
+- It is now validated that the filter operator matches the field type.
 - Allow passing page as string to `Flop.set_page/2`.
 - Allow passing offset as string to `Flop.set_offset/2`.